chore(oonimkall): re-enable TestTaskRunnerRun
I have ran the test several times locally with `go test -race -count 1`
and it was working as intended. Either the CI run issue was just temporary
or it's something fundamental of the CI that I can'r reproduce.

For this reason, I am going to re-enable the test and see what happens.

Closes https://github.com/ooni/probe/issues/2541