feat: mvp of conditionally enabling experiments
We have a class of experiments that I call experimental experiments.

As a reminder, experimental experiments are the ones we have just
added in a development cycle and which may potentially have issues
down the line, so that we may want to disable them while working
on an emergency release to fix them.

Currently, riseupvpn and echechk fall into this class.

The idea of this diff is to implement the functional specification
described by issue https://github.com/ooni/probe/issues/2553.