[backport] fix(registry): mark torsf as disabled by default
This diff backports #1359 to the release/3.19 branch.

There may be upcoming changes in torsf which may cause it to fail
consistently as it occurred during the 3.18 cycle.

Therefore, be defensive and make it disabled by default.

Part of https://github.com/ooni/probe/issues/2553

While there, use slightly better naming for an echcheck function.