[backport] refactor(riseupvpn): rename structs, change progress, bump version
This diff backports #1361 to the release/3.19 branch.

This diff renames some structs, changes the code emitting the progress,
and bumps the experiment version. In my previous commit, I said I did
bump the version number but that was not actually the case.

This diff has been extracted from
https://github.com/ooni/probe-cli/pull/1125.

This diff is part of https://github.com/ooni/probe/issues/1432

---------

Co-authored-by: cyBerta <cyberta@riseup.net>