feat(riseupvpn): include bootstrap in progress
This diff modifies riseupvpn to include bootstrap in progress.

We define as riseupvpn's bootstrap the process of fetching the CA and
the various JSON files needed to find out the gateways to measure.

I previously stopped including bootstrap in progress in
https://github.com/ooni/probe-cli/pull/1363, which I did
because the progress was not monotonic.

This diff introduces helpers that make it very obvious to emit
monotonic bootstrap through the ./internal/progress package.

This work is part of https://github.com/ooni/probe/issues/1432.