refactor(buildtool): use list for OpenSSL ./Configure flags (#1367)
We currently have a string variable in `cBuildEnv` named
`OPENSSL_API_DEFINE` that we append to OpenSSL's `./Configure`
invocation to force using the proper Android API. However, for building
for iOS (a need documented by
https://github.com/ooni/probe/issues/2564), we need a list of strings,
because there is more than a single scalar that we need to append to the
`./Configure` invocation (as shown by the MVP implementation at
https://github.com/ooni/probe-cli/pull/1366).

Hence, this diff, which introduces a string list named
`OPENSSL_POST_COMPILER_FLAGS` that contains strings to append to the
`./Configure` command line _after_ the OS/compiler flag. We specifically
named the variable "post compiler" because there is another variable in
the same `cBuildEnv` struct called `OPENSSL_COMPILER`.