[backport] chore(buildtool): add ios-build unit tests
This diff backports #1371 to the release/3.19 branch.

We needed to move the runtimex.Assert for darwin, otherwise we cannot
run unit tests on linux 😅. While there, be consistent and make sure we
avoid panicking on Windows if one runs unit tests there.

Part of https://github.com/ooni/probe/issues/2564