refactor(dslx): Runtime is a renamed, abstract ConnPool
We want to use the DSL inside the oohelperd. We don't care about
collecting observations in the oohelperd. So, the plan is that of
abstracting the ConnPool, renaming it Runtime, and giving it the
power to create abstract Traces.

The oohelperd will use a MinimalRuntime (sketched out by this
commit) that will not collect any observation.

Measuring code, instead, will use a MeasurexRuntime that will
collect observations.

This commit is just the first step. We rename and introduce the
MinimalRuntime. No significant functional changes so far.