refactor(dslx): use an abstract trace
This diff builds on the previous diff and uses an abstract trace
inside of the dslx package. By using an abstract trace, we can
choose between using:

- a runtime that collects observations, based on measurexlite; and

- a minimal runtime that does not collect observations.

To make the trace abstract, we need to modify measurexlite's trace
such that it can be used as an interface.

In turn, this means we need to update the throttling package
such that it uses an abstract trace definition. Strictly speaking,
we could have avoided introducing this abstraction, but it seems
better to also use an abstract trace there, as it allows for
improving the decoupling with measurexlite.