refactor(dslx): Runtime includes logger, ID generator, and zero time
Currently, we pass these fields to each DSL function. However, if
we want to load the functions from JSON, which is something we have
experimented with in the richer input context, we can't do this.

These fields should instead belong to the runtime. A subsequent diff
will modify the DSL functions to take them from the runtime.