refactor(dslx): prepare for improving testing
Rather than using custom fields for testing, we can configure in the
runtime a custom model.MeasuringNetwork.

We're not doing this just to simplify the codebase, rather the underlying
intent here is making sure we don't need to keep much state in each
function, so we can refactor them to be pure functions wrapped by an
adapter that produces the desired type.

In turn, by doing that, we will be able to factor complexity around
invoking functions and parsing their results.

In turn, by doing that, we will be able to modify the signature of the
functions and do the following:

1. allow the DSL model to include stages that take in input a Maybe
value rather than a value, so we can observe failures more easily than
we do now and we can write inline code to save into test keys;

2. allow the DSL model to much more easily be refactored to use
channels, which in turn enables us to compose operations more naturally
and increase the amount of overlapping (think, e.g., how this enables
the possibility of waiting additional time for a DNS-over-UDP resolver
to wait for late/duplicate replies).