refactor(dslx): start using pure functions
Introduce an adapter type that converts a pure function into a Func
and start using it whenever it's easy, rather than rolling out structs
that implement the Func type when we actually don't have state.

This change is an improvement because we're creating the necessary
conditions from moving complexity out of the functions that actually
do something, which should be simpler, and adapters, which should
contain the same, equal logic for creating pipelines.

The ultimate goal here is to be able to have stages that accept
a Maybe[A] as input, to be able to add inline evaluators that set
test keys. I want to reach this goal by making the necessary
transformation at the adapters level rather than changing each
of the Func prototype at once.