refactor(dslx): reduce tlsHandshakeFunc state
This diff reduces the state kept by the tlsHandshakeFunc struct
so that we can apply a transformation similar to the one we applied
for TCPConnect() and implement TLSHandshake() using a pure func.

The overall objective is that of factoring away completixity to
enable manipulating this code more easily.

While there, let's note that the changes applied here mean that we
can reuse this code for configuring tls.Config for the QUICHandshake.