fix(nettests): failing to start a nettest is a warning (#1373)
Emitting an error causes ooni/probe-desktop to try to generate a warning
window that (1) is a bit unpleasant to visually see and (2) more
importantly is not proper to show, given that we have default-disabled
experiments right now.

Also, in my book, we should emit warning if we continue running and
errors when we stop. (Maybe it's not the best book considering that
assumes that all errors are fatal; in any case, try to be consistent
with this convention.)

See https://github.com/ooni/probe/issues/2576 for extra information.