[backport] fix(registry): disable vanilla_tor by default
This diff backports #1374 to release/3.19.

The reference issue is https://github.com/ooni/probe/issues/2553.

When I implemented the disabled-by-default functionality, I missed that
we should also disable-by-default vanilla_tor.

The reason for doing that is that the experiment crashes on Android and
possibly iOS due to https://github.com/ooni/probe/issues/2406.