break the code in a different way
Because I am dropping the requests again, we break again the tests
with the redirects. I could possibly fix it by putting requests back
again but I am not super happy about doing this because that would
cause the DSL to do some strange work and I'd honestly rather not do this.