feat: qatool to run webconnectivity lte QA (#1394)
This diff introduces the qatool, originally developed in
https://github.com/ooni/probe-cli/pull/1392. This tool allows to run
webconnectivitylte netemx-based QA tests and produces the measurement,
observations, and analysis files.

We can use this tool to run individual tests and inspect their results
when they're failing. Additionally, we can use this tool to regenerate
the test suite used by the minipipeline.

When running the tool, we can select which QA tests to run, whether to
rerun netemx and to reprocess. This feature is handy because we don't
need to rerun netemx every time and sometimes we may only want to rerun
netemx.

This work is part of https://github.com/ooni/probe/issues/2634. It helps
us to manage and test the minipipeline and webconnectivitylte in an
easier way, and it enables further development and testing.