fix(minipipeline/analysis): distinguish between None and empty (#1401)
None means that an algorithm did not run or did not find enough data to
produce a result. Empty means that the algorithm did run, did find
enough data, and did produce an empty result.

The difference between these two states, which generally is not
important when writing Go code, is extremely important to take the
correct decisions when assigning the results of web measurements.

Accordingly, this diff goes through each algorithm and ensures we start
with a None state and only switch to the empty state when we have seen
enough data to determine that the result is indeed empty.

Additionally, in this diff we also add the following new analysis rules:

- `ComputeDNSPossiblyInvalidAddrsClassic`, which is like
`ComputeDNSPossiblyInvalidAddrs` but does not consider TLS, which in
turn is useful to emulate the original Web Connectivity v0.4 behavior;
- `ComputeDNSPossiblyNonexistingDomains`, which tells us for which
domains the probe and the control agree that those domains are undefined
(i.e., they both get `NXDOMAIN`), which is useful to detect this
specific case;

We also renamed `ComputeHTTPFinalResponses` to
`ComputeHTTPFinalResponsesWithControl` and changed the definition such
that we only include responses for which we have a control. This is the
core rule to decide whether we should move forward with considering the
results of the HTTP diff set of algorithms.

The reference issue is https://github.com/ooni/probe/issues/2634.