fix(minipipeline): change DNSDiff algorithm (#1407)
We were running the algorithm on single IP addresses resolved by the
probe against the set resolved by the control.

This is incorrect, because we should compare to the whole set resolved
by the probe. This fact is relevant because v0.4 says there's
consistency if a single IP address or ASN resolved by the probe
intersects with the corresponding control set.

Conversely, our comparison required that each IP address resolved by the
probe belonged to the control set of addresses or ASNs.

This diff solves the problem.

While there, remove DNSPossiblyInvalidAddrs. I have determined we need
to run DNSDiff independently of bogons checks.

Incidentally, the new DNSDiff code I am adding is straight from Web
Connectivity v0.4, so we should be good.

Part of https://github.com/ooni/probe/issues/2634