feat(minipipeline): add metrics required by LTE compat analysis mode (#1415)
This diff implements the following useful metrics:

* DNSLookupSuccessWithValidAddress
* DNSLookupSuccessWithValidAddressClassic
* DNSLookupExpectedFailure
* DNSLookupExpectedSuccess
* HTTPFinalResponseSuccessTCPWithoutControl

It also renames the HTTPDiff metrics for consistency, by adding an
HTTPFinalResponse prefix to names.

It also reimplements DNSExperimentFailure in terms of a generic DNS data
collection function. And it also moves the definition of this variable
to be near the top of the structure.

I have also removed some tests. While in general this goes against my
"""religion""", I think I had added tests too prematurely and I need to
move a bit faster to make this work converge. (Also, there are still
several many tests that deal with producing data, which is why we have
so much churn at every diff that changes the minipipeline. Also, the
tests I am removing were only to cover corner cases, so I think it's
overall justifiable to do this.)

Part of https://github.com/ooni/probe/issues/2634