fix(oonimkall): make sure SoftwareName is set (#1425)
While there, make sure we log the check-in request and response with
`-v`, because otherwise 🤌🤌🤌🤌. (On this note, we did not want to log
psiphon and tor configs, but we're not using check-in for them anymore,
so this switch from not logging to logging seems indeed to be reasonably
okay to do. We need /api/v1/register and /api/v1/login before serving
config. Additionally, it seems having experiment-specific APIs is
actually more flexible anyway.)

Part of https://github.com/ooni/probe/issues/2644.