feat(minipipeline): add ControlFinalResponseExpectations (#1451)
This PR adds the ControlFinalResponseExpectations field to
WebObservationsContainer. We need this field to determine whether
unexplained failures observed by the probe (i.e., failures occurring
during redirects) are to be expected (because for some reason also the
TH failed) or unexpected (because the TH succeeded).

Part of https://github.com/ooni/probe/issues/2640.