feat(webconnectivitylte): handle ghost DNS censorship
This commit modifies webconnectivitylte to handle ghost DNS censorship.

We define ghost DNS censorship the case where the original domain does not
exist anymore but the censor continues to return an IP address for the original
domain nonetheless.

We used to have null-null handling for this case in the "orig" engine and
a reference issue as https://github.com/ooni/probe/issues/2307.

With this commit, we modify the "classic" engine to correctly handle this case.

In doing this, we notice that we should probably start departing from strict
v0.4 orthodoxy and instead produce better results in the following cases:

1. mark DNS inconsistency when we have successful probe lookups and no
IP address has been resolved by the test helper;

2. specific handling on the case in which the website seems down where we
also ask ourselves the question of whether the culprit could be the DNS.

These two changes are required to handle ghost DNS censorship.

Additionally, let's recognize that always setting HTTPExprimentFailure is
probably for the greater good anyway, given that we're stopping to be very
compatible with v0.4 anyway.

Part of https://github.com/ooni/probe/issues/2652.