fix(minipipeline): make expected TCP, TLS failures WAI
This diff corrects an embarassing bug in the logic we were using
for expected TCP connect and TLS handshakes failures.

We spotted this issue with the new testcase we added called
websiteDownTCPConnect, which is also included.

Additionally, we needed to fix some tests because the logic bugs
were preventing us from seeing some events.

Part of https://github.com/ooni/probe/issues/2652.

Closes https://github.com/ooni/probe/issues/2299.