chore(minipipeline): add test case for ooni/probe#2456 (#1461)
This commit adds a test case that replicates the issue described by
https://github.com/ooni/probe/issues/2456 and shows that we are now able
to correctly deal with that. Check the `analysis.json` and
`analysis_classic.json` file and see that there are no unexplained or
unexpected TCP connect failures. This happen because now we filter for
IPv6-related failures.

See also
https://explorer.ooni.org/m/20240123143658.130691_IT_webconnectivity_03662f69246c637f.

Closes https://github.com/ooni/probe/issues/2456.