fix(oohelperd): make sure endpoints don't connect to 127.0.0.1 (#1463)
This diff adds extra tests and logging showing that we're not connecting
to 127.0.0.1 in endpoints code. We still need to deal with making sure
we don't connect to 127.0.0.1 in HTTP code. In principle, this is
already true, but the QA framework constructs the oohelperd differently
than with the default constructor, and we'll need to improve in this
department to make sure the results obtained in the QA suite also give
us confidence on the general oohelperd behavior. For now, QA tests
attempting to use 127.0.0.1 do not produce the correct result because
oohelperd misbehaves.

Extracted from https://github.com/ooni/probe-cli/pull/1462.

Part of https://github.com/ooni/probe/issues/2652.

Part of https://github.com/ooni/probe/issues/1517.