cleanup(netxlite): remove first-order implicit-Netx wrappers (#1465)
This diff removes the first-order implicit-Netx wrappers. We define
first-order implicit-Netx wrappers the top-level functions that have the
same name of netxlite.Netx methods, allocate an empty Netx, and call the
corresponding method.

The reason why we're doing this now is that it has been relatively hard
to implement https://github.com/ooni/probe-cli/pull/1464 because of the
ambiguity between those first-order wrappers and the methods. Getting
this wrong means that QA tests would behave in a funny way.

The overall goal here is to refactor `oohelperd` to only depend on
`netxlite.Netx` such that we can remove the code duplication between how
we instantiate `oohelperd.Handler` in `oohelperd` and how we instantiate
it inside `netemx`.

In turn, by doing this, we would ensure we have the same `oohelperd`
behavior for QA and production.

In turn, with this guarantee, we can write QA tests that ensure we're
correctly dealing with 127.0.0.1.

The reference issue is https://github.com/ooni/probe/issues/1517.