refactor(oohelperd): depend on netxlite.Netx only
This diff modifies netxlite such that the oohelperd only depends
on the netxlite.Netx instance.

The overall goal here is to refactor `oohelperd` to only depend on
`netxlite.Netx` such that we can remove the code duplication between how
we instantiate `oohelperd.Handler` in `oohelperd` and how we instantiate
it inside `netemx`.

In turn, by doing this, we would ensure we have the same `oohelperd`
behavior for QA and production.

In turn, with this guarantee, we can write QA tests that ensure we're
correctly dealing with 127.0.0.1.

The reference issue is https://github.com/ooni/probe/issues/1517.