refactor(oohelperd,netmx): reduce construction diff to zero (#1467)
This diff modifies how we construct oohelperd.Handler inside the
oohelperd package and inside the netemx package such that we're now
constructing using equivalent code. We know that construction is
equivalent for HTTP clients because previosuly we made sure it was the
case in https://github.com/ooni/probe-cli/pull/1464.

So, the next step would be removing the custom construction code inside
of netemx and always use oohelperd.NewHandler.

In turn, by doing this, we would ensure we have the same `oohelperd`
behavior for QA and production.

In turn, with this guarantee, we can write QA tests that ensure we're
correctly dealing with 127.0.0.1.

The reference issue is https://github.com/ooni/probe/issues/1517.