fix(webconnectivitylte): handle measurements with loopback addrs (#1462)
This diff modifies webconnectivitylte and related packages to correctly
handle measurements containing loopback addresses. There are two cases:
(a) when both the probe and the TH only see loopback addresses at least
for the getaddrinfo lookups; (b) when only the probe sees loopback
addresses at least for the getaddrinfo lookups. In the former case, we
should mark the website as "down" (maybe a little bit of a stretch, but
if we mark down a website where the TLS is misconfigured, arguably we
can do the same when DNS is misconfigured). In the latter case, we mark
the measurement as censorship, since seeing a loopback address is not
what we expect.

Closes https://github.com/ooni/probe/issues/1517.