fix(webconnectivitylte): handle domains w/o A/AAAA records (#1473)
This diff improves webconnectivitylte and minipipeline to correctly
handle the case where a website has no configured IP addresses but the
domain for the website exists. When this happens, the TH returns no
error but an empty list of resolved addresses for historical reasons,
while the probe fails with the `"dns_no_answer"` error. We need to
intercept and correctly handle this scenario both for the ordinary
analysis and the classic analysis.

Let's also add QA test cases to make sure we don't regress.

Part of https://github.com/ooni/probe/issues/2652.