fix(webconnectivitylte): handle malformed redirect URLs
This diff ensures that webconnectivitylte is able to handle malformed
redirect URLs such as (literally) `http://` and `https://`.

The way in which we do this is slightly different from v0.4 and possibly
more accurate in that it attributes the error to the operation where we
detect the error rather than later on in the next redirect.

Because of that, I made the QA suite conform to v0.5's behavior.

Closes https://github.com/ooni/probe/issues/2628