fix(script/go.bash): support Ubuntu 22.04 LTS and Debian 11 (#1487)
This commit is part of https://github.com/ooni/probe/issues/2664. We use
a different strategy for downloading and executing the correct version
of Go. We have vendored and adapted to code at `golang.org/dl` such that
it works with versions of Go as old as the version of Go used by Ubuntu
22.04 LTS and Debian 11.