BREAKING CHANGE: feat: move StreamAllContext to netxlite (#1490)
This diff moves the `StreamAllContext` implementation from
`webconnectivitylte` to `netxlite`.

We flagged this diff as BREAKING CHANGE because we also fixed a bug
where, in case of context timeout, `StreamAllContext` was suppressing
the error rather than reporting it. Re-reading the code, that felt
incorrect because it did not allow us to clearly know that we timed out
reading the body, which could be caused by throttling.

The new behavior seems more accurate.

Because of this change, I also felt it was time to add explicit tests
for cases where we download fails because it takes too much time for us
to fetch the whole response body. I am not 100% sure we are correctly
flagging this case, but an `http-failure` probably seems fine at the
moment and we can always revisit this as we learn more about throttling.

Part of https://github.com/ooni/probe/issues/2654.