fix(gardener): avoid issues caused by stale databases
We use a database to keep measurements results to allow to interrupt
and resume the work, given that we have thousands or URLs to vet.

However, after a week, the database is most likely stale, so let's
just avoid trusting and start afresh in such a case.

It's reasonable to assume that the database is stale after such an
interval because the URL's status may have changed (e.g., some URLs
could have stopped working).

Additionally, every time we run `gardener sync` we most likely
have new URLs, so we need to do one of the following:

1. merge new URLs into the database;

2. just zap the database and start afresh.

Because doing 1. would be a bit time consuming, for now I am opting
for doing 2. and we'll implement 1. if we really need it.