cleanup(engine): do not expose LookupLocationContext
We can obtain the ~same effect by using MaybeLookupLocationContext
and then using the corresponding Session accessors.

Potentially, this code would cause a behavioral change in that
once the probe location has been found, it would not change again
and it might be a problem for very-long-running sessions.

However:

1. the Android codebase does not keep a reference to a session
for a very long time and anyway the longest-running sessions are
those used for running experiments;

2. the correct behavior would be for MaybeLookupLocationContext
to cache the results only for a limited amount of time.

Because of all these considerations, it actually makes sense
to say that "we're obtaining the ~same effect".

The net benefit for us is that we can further reduce the surface
of interaction between clients and the engine code. A simpler
API surface is also simpler to document.

Part of https://github.com/ooni/probe/issues/2700