cleanup: SaveMeasurement doesn't belong to Experiment
This cleanup may be a bit tangential, but it still makes sense in
the context of what I'm doing, i.e., cleaning up the engine API and
ensuring we mostly use engine functionality in clients.

While doing this, it occurred to me that I could avoid keeping the
SaveMeasurement method attached to an experiment, and I could instead
just make it a pure function.

This reduces the complexity of the *engine.Experiment type.

While there, slightly improve testing.

Part of https://github.com/ooni/probe/issues/2700